// src/app/models/assetinfo.model.ts
export class Asset {
    id?: number = 0;
    assetname:  string = '';
    serialnumber:  string = '';
    ponumber:  string = '';
    assettype: string = '';
    statusofasset:  string = '';
    projectname:  string = '';
    pmname:  string = '';
    devicemanager:  string = '';
    user:  string = '';
    location:  string = '';
    email:  string = '';
  }
  